fix (all products): rename products sub-reducer and use combineReducers properly
