+++ client/store/index.js
@@ -3,9 +3,9 @@ import createLogger from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
 import user from './user'
-import product from './product'
+import products from './products'
 
-const reducer = combineReducers(user, product)
+const reducer = combineReducers({user, products})
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
 )
@@ -13,4 +13,4 @@ const store = createStore(reducer, middleware)
 
 export default store
 export * from './user'
-export * from './product'
+export * from './products'
